- Changed tree structure of the project. Instead of having 'include' folder for headers and 'src' folder for sources I propose to use module folders inside 'src'. Look inside for a better comprehension. - Refactored Potential, splitting the builder implementation into PotentialBuilder.cpp - Created Base and DiscreteBase ( work in progress, it's not yet how it should be! ) - Changed CMakeLists.txt in order to compile on Windows. Need some tests...
