+++ src/Potential/Potential.h
@@ -1,5 +1,5 @@
-#ifndef POTENTIAL
-#define POTENTIAL
+#ifndef POTENTIAL_H
+#define POTENTIAL_H
 
 #include <iostream>
 #include <vector>
@@ -52,46 +52,13 @@ public:
             double height        = 10.0;
 
         public:
-            Builder(std::vector<double> x_new) {
-                this->x = x_new;
-            }
-
-            Builder setK(double k_new){
-                this->k = k_new;
-                return *this;
-            }
-
-            Builder setWidth(double width_new){
-                if (width_new >= 0) {
-                    this->width = width_new;
-                    return *this;
-                }
-                else throw std::invalid_argument("Width parameter cannot be negative.");
-            }
-
-            Builder setHeight(double height_new){
-                this->height = height_new;
-                return *this;
-            }
-
-            Builder setType(std::string type){
-                if (!type.empty()) {
-                    this->type = type;
-                    return *this;
-                }
-                else throw std::invalid_argument("Empty type given as parameter.");
-            }
-
-            Potential build(){
-                try {
-                    return Potential(this->x,this->type,this->k,this->width,this->height);
-                }
-                catch(const std::invalid_argument& e){
-                    throw;
-                }
-            }
+            Builder(std::vector<double> x_new);
+            Builder setK(double k_new);
+            Builder setWidth(double width_new);
+            Builder setHeight(double height_new);
+            Builder setType(std::string type);
+            Potential build();
     };
 };
-
 
 #endif
